# docs/openapi.yaml
openapi: 3.1.0
info:
  title: RunPod ComfyUI Worker API
  version: 0.1.0
  description: >-
    Reference OpenAPI definition for the local FastAPI wrapper around the RunPod
    ComfyUI worker. Use this file with Swagger UI or any OpenAPI viewer to
    inspect request and response payloads.
servers:
  - url: http://localhost:3000
    description: Local development server
paths:
  /run:
    post:
      summary: Submit an asynchronous ComfyUI workflow
      description: >-
        Wraps a workflow template and forwards it to the ComfyUI API. Returns a
        generated job identifier that can be polled via the `/status/{job_id}`
        endpoint. Supported `input.comfyui_workflow_name` values are documented
        at `GET /docs/workflow-templates`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RunRequest'
            example:
              input:
                prompt: "A cinematic slow zoom across a neon city"
                image: "iVBORw0KGgoAAAANSUhEUg..."
                width: 480
                height: 640
                length: 81
                comfyui_workflow_name: "video_wan2_2_14B_i2v"
      responses:
        '200':
          description: Job accepted and queued for processing.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RunResponse'
              example:
                id: "42be01db-ea96-494a-a156-94c5857d865c"
                status: "QUEUED"
        '400':
          description: Invalid request payload or workflow configuration.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '502':
          description: Upstream ComfyUI service error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /status/{job_id}:
    get:
      summary: Retrieve asynchronous job status
      description: >-
        Returns the current status and any generated assets for a previously
        submitted workflow.
      parameters:
        - name: job_id
          in: path
          required: true
          description: Identifier obtained from the `/run` response.
          schema:
            type: string
      responses:
        '200':
          description: Current job state. Includes asset payloads when completed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusResponse'
              examples:
                videoCompleted:
                  summary: Completed video generation job
                  value:
                    delayTime: 50435
                    executionTime: 258789
                    id: "42be01db-ea96-494a-a156-94c5857d865c-e2"
                    output:
                      output:
                        images: []
                        videos:
                          - data: "AAAAIGZ0eX.............."
                            filename: "ComfyUI_00001_.mp4"
                            type: "base64"
                    status: "COMPLETED"
                    workerId: "h721wnkh45pang"
        '404':
          description: Job identifier is unknown to the worker.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Unexpected worker error when resolving job status.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  schemas:
    RunRequest:
      type: object
      required:
        - input
      properties:
        input:
          type: object
          description: >-
            RunPod job payload that the worker forwards to ComfyUI. Fields differ
            per workflow; the listed attributes are used by the video
            `video_wan2_2_14B_i2v` template.
          properties:
            prompt:
              type: string
              description: Natural language description that guides the workflow.
            image:
              type: string
              format: byte
              description: Base64-encoded input image (PNG or JPEG recommended).
            width:
              type: integer
              minimum: 64
              default: 480
              description: Output video width in pixels.
            height:
              type: integer
              minimum: 64
              default: 640
              description: Output video height in pixels.
            length:
              type: integer
              minimum: 1
              default: 81
              description: Number of frames to generate.
            comfyui_workflow_name:
              type: string
              default: video_wan2_2_14B_i2v
              description: Identifier of the workflow template to load.
            comfy_org_api_key:
              type: string
              nullable: true
              description: Optional Comfy.Org API key.
    RunResponse:
      type: object
      required:
        - id
        - status
      properties:
        id:
          type: string
          format: uuid
          description: Generated identifier for the asynchronous job.
        status:
          type: string
          enum: [QUEUED]
          description: Initial status after the job is accepted.
    StatusResponse:
      type: object
      required:
        - id
        - status
      properties:
        id:
          type: string
          description: Job identifier provided by the `/run` endpoint.
        status:
          type: string
          description: Uppercase job lifecycle state.
          enum: [QUEUED, RUNNING, COMPLETED, FAILED]
        created_at:
          type: number
          format: double
          nullable: true
          description: Timestamp (seconds since epoch) when the job was queued locally.
        updated_at:
          type: number
          format: double
          nullable: true
          description: Timestamp of the last state change tracked by the worker.
        delayTime:
          type: integer
          format: int64
          nullable: true
          description: Milliseconds spent in the RunPod queue before execution (if available).
        executionTime:
          type: integer
          format: int64
          nullable: true
          description: Milliseconds spent executing on the assigned worker (if available).
        workerId:
          type: string
          nullable: true
          description: Identifier of the RunPod worker that processed the job.
        error:
          type: string
          nullable: true
          description: Present when `status` is `FAILED`.
        output:
          type: object
          nullable: true
          description: Generated assets packaged by the worker.
          properties:
            output:
              $ref: '#/components/schemas/WorkerOutput'
    WorkerOutput:
      type: object
      properties:
        images:
          type: array
          items:
            $ref: '#/components/schemas/Asset'
        videos:
          type: array
          items:
            $ref: '#/components/schemas/Asset'
    Asset:
      type: object
      required:
        - filename
        - type
        - data
      properties:
        filename:
          type: string
          description: Asset name returned by ComfyUI.
        type:
          type: string
          description: Serialization method for `data` (e.g., `base64`, `s3_url`).
        data:
          type: string
          description: Encoded payload or remote reference, depending on `type`.
    ErrorResponse:
      type: object
      required:
        - detail
      properties:
        detail:
          type: string
          description: Human-readable error message.